+++ src/containers/ScoreBoard/styles.ts
@@ -1,23 +1,41 @@
 import styled, { css } from "styled-components";
 
-const Wrapper = styled.div`
-  max-width: 500px;
-  margin: 0 auto;
-  padding: 15px;
-  background: #383838;
-  border-radius: 5px;
+const Bottom = styled.div`
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
 `;
 
-interface SectionProps {
+const ScoreBoard = styled.div`
+  /* max-width: 500px; */
+  /* margin: 0 auto; */
+  padding: 15px;
+  background: #383838;
+  /* border-radius: 5px; */
+
+  display: flex;
+  justify-content: space-between;
+`;
+
+const Column = styled.div`
+  width: 100%;
+  padding: 0 15px;
+`;
+
+const Left = styled(Column)``;
+const Right = styled(Column)``;
+
+export interface RowProps {
   col?: boolean;
 }
 
-const Section = styled.div`
+const Row = styled.div<RowProps>`
   display: flex;
   justify-content: space-between;
   margin-bottom: 15px;
 
-  ${(props: SectionProps) =>
+  ${(props) =>
     props.col &&
     css`
       flex-direction: column;
@@ -25,16 +43,16 @@ const Section = styled.div`
     `}
 `;
 
-interface TextProps {
+export interface TextProps {
   padBottom?: boolean;
 }
 
-const Text = styled.span`
+const Text = styled.span<TextProps>`
   font-weight: 300;
   color: #f8f8f8;
   font-family: "Roboto", sans-serif;
 
-  ${(props: TextProps) =>
+  ${(props) =>
     props.padBottom &&
     css`
       padding-bottom: 10px;
@@ -46,4 +64,4 @@ const BoldText = styled(Text)`
   color: #f8f8f8;
 `;
 
-export { Wrapper, Text, BoldText, Section };
+export { ScoreBoard, Text, BoldText, Row, Left, Right, Bottom };