+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- random event foundational logic
+- build modal enhanced
+- yearly advance foundational logic
+
+## Added
+
+- player data (name, ID) are now cached
+
 ## [0.2.0] - 2019-05-25
 
 ### Added